refactor: 소식탭 뷰어들에 suspense 적용
