+++ app/[locale]/community/notice/[id]/page.tsx
@@ -1,24 +1,15 @@
-import { getNoticePostDetail } from '@/apis/notice';
+import { Suspense } from 'react';
 
-import Attachments from '@/components/common/Attachments';
-import { StraightNode } from '@/components/common/Nodes';
-import Tags from '@/components/common/Tags';
-import HTMLViewer from '@/components/editor/HTMLViewer';
-import PageLayout, { PAGE_PADDING_BOTTOM_PX } from '@/components/layout/pageLayout/PageLayout';
-import PostFooter from '@/components/post/PostFooter';
+import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
 import { PostSearchQueryParams } from '@/types/post';
 
-import { formatPostDateStr } from '@/utils/date';
-import { getPath } from '@/utils/page';
-import { notice } from '@/utils/segmentNode';
+import NoticeViewer from './NoticeViewer';
 
 interface NoticePostPageProps {
   params: { id: string };
   searchParams: PostSearchQueryParams;
 }
-
-const noticePath = getPath(notice);
 
 export default async function NoticePostPage({
   params: { id: rawID },
@@ -29,23 +20,11 @@ export default async function NoticePostPage({
   // ID가 잘못된 경우 예외 처리
   if (Number.isNaN(id)) return <InvalidIDFallback rawID={rawID} />;
 
-  const notice = await getNoticePostDetail(id, searchParams);
-
   return (
     <PageLayout titleType="big" bodyStyle={{ padding: 0 }}>
-      <Header {...notice} />
-      <div
-        className="bg-neutral-50 px-5 pt-9 sm:pl-[100px] sm:pr-[340px]"
-        style={{
-          paddingBottom: PAGE_PADDING_BOTTOM_PX,
-        }}
-      >
-        <Attachments files={notice.attachments} />
-        <HTMLViewer htmlContent={notice.description} className="mb-10" />
-        <StraightNode />
-        <Tags tags={notice.tags} margin="mt-3 ml-6" searchPath={noticePath} />
-        <PostFooter post={notice} postType="notice" id={rawID} margin="mt-12" />
-      </div>
+      <Suspense>
+        <NoticeViewer id={id} searchParams={searchParams} />
+      </Suspense>
     </PageLayout>
   );
 }
@@ -57,23 +36,3 @@ const InvalidIDFallback = ({ rawID }: { rawID: string }) => (
     </p>
   </PageLayout>
 );
-
-const Header = ({
-  title,
-  author,
-  createdAt,
-}: {
-  title: string;
-  author: string;
-  createdAt: string;
-}) => {
-  return (
-    <div className="flex flex-col gap-4 px-5 py-9 sm:pl-[100px] sm:pr-[340px]">
-      <h2 className="text-[1.25rem] font-semibold">{title}</h2>
-      <div className="flex gap-5 text-sm font-normal tracking-wide text-neutral-500">
-        <p>글쓴이: {author}</p>
-        <p>작성 시각: {formatPostDateStr(createdAt)}</p>
-      </div>
-    </div>
-  );
-};
